doit: Split out cmd_replace
